+++ src/App.js
@@ -1,6 +1,10 @@
 import React, { Component } from 'react';
+import { Route, Routes } from 'react-router-dom';
+import Header from './components/Header';
+import Calculator from './components/pages/calculator';
+import Home from './components/pages/home';
+import Quote from './components/pages/quote';
 import './App.css';
-import Calculator from './components/calculator';
 
 class App extends Component {
   constructor(props) {
@@ -10,7 +14,14 @@ class App extends Component {
 
   render() {
     return (
-      <Calculator />
+      <div>
+        <Header />
+        <Routes>
+          <Route path="/" element={<Home />} />
+          <Route path="/calculator" element={<Calculator />} />
+          <Route path="/quote" element={<Quote />} />
+        </Routes>
+      </div>
     );
   }
 }
