+++ src/components/pages/home.js
@@ -1,0 +1,21 @@
+import React from 'react';
+
+const Home = () => (
+  <>
+    <h2>Welcome to our page</h2>
+    <p>
+      Lorem ipsum, dolor sit amet consectetur adipisicing elit.
+      Quas magni ab expedita natus dolor architecto, ducimus
+      reiciendis voluptate consequatur cumque, maiores iusto temporibus,
+      tempora reprehenderit autem. Officia culpa eligendi cumque.
+    </p>
+    <p>
+      Lorem ipsum, dolor sit amet consectetur adipisicing elit.
+      Quas magni ab expedita natus dolor architecto, ducimus
+      reiciendis voluptate consequatur cumque, maiores iusto temporibus,
+      tempora reprehenderit autem. Officia culpa eligendi cumque.
+    </p>
+  </>
+);
+
+export default Home;
