+++ src/components/pages/quote.js
@@ -1,0 +1,10 @@
+import React from 'react';
+
+const Quote = () => (
+  <p className="quote">
+    Mathematics is not about numbers, equations, computations, or algorithms:
+    it is about understanding. -William Paul Thurston
+  </p>
+);
+
+export default Quote;
